requirements.txt
altgraph==0.17.4
attrs==23.2.0
auto-py-to-exe==2.44.1
beautifulsoup4==4.10.0
bottle==0.12.25
bottle-websocket==0.2.9
bs4==0.0.2
certifi==2024.7.4
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.3.2
chromedriver-autoinstaller==0.6.4
clr-loader==0.2.6
cryptography==42.0.8
dbfread==2.0.7
Eel==0.17.0
exceptiongroup==1.2.2
fake-useragent==1.5.1
future==1.0.0
gevent==24.2.1
gevent-websocket==0.10.1
greenlet==3.0.3
h11==0.14.0
idna==3.7
numpy==2.0.1
outcome==1.3.0.post0
packaging==24.1
pandas==2.2.2
pefile==2023.2.7
pillow==10.4.0
proxy_tools==0.1.0
pycparser==2.22
pyinstaller==6.9.0
pyinstaller-hooks-contrib==2024.7
pynput==1.7.7
pyparsing==3.1.2
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pythonnet==3.0.3
pytz==2024.1
pywebview==5.1
pywin32-ctypes==0.2.2
requests==2.32.3
schedule==1.2.2
selenium==4.23.1
setuptools==72.1.0
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.5
trio==0.26.0
trio-websocket==0.11.1
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
websocket-client==1.8.0
whichcraft==0.6.1
wsproto==1.2.0
wxPython==4.2.1
xlrd==2.0.1
zope.event==5.0
zope.interface==6.4.post2
